fix(ci): unblock builds on double-promotion, nodiscard, and Perl packaging
Cast percentile math to long double, honor wait_for_active_operations
return value in backup_to, and drop ripgrep from package-perl-client.sh
so hosted runners without rg can package the SDK.

+++ include/glyphastore/core/latency_histogram.hpp
@@ -59,7 +59,8 @@ struct LatencyHistogram final {
         if (observations == 0 || !(percentile >= 0.0) || percentile > 1.0) {
             return 0;
         }
-        const auto target = static_cast<std::uint64_t>(static_cast<long double>(observations) * percentile);
+        const auto target = static_cast<std::uint64_t>(static_cast<long double>(observations) *
+                                                       static_cast<long double>(percentile));
         const auto rank = target == 0 ? 1U : std::min(target, observations);
         const auto cumulative = this->cumulative();
         std::uint64_t previous = 0;
